Added: Tests for symmetric compression (from pdf)
